requirements.txt
fastapi
uvicorn
ccxt==2.1.78
python-telegram-bot==13.13
pandas==1.3.4
numpy==1.26.3
websockets
gunicorn
python-jose
passlib[bcrypt]
redis
celery
aiohttp
sqlalchemy
pydantic